requirements.txt
requests==2.31.0
Flask==3.0.3
Flask-Cors==4.0.1
pyngrok==7.1.6
python-dotenv==1.0.1
pytube==15.0.0
keyboard==0.13.5